Fix handling of creating projects when not also creating a virtualenv
Previously the `init` command assumed the project was being setup
within the root of the virtualenv. The assumption now is that the
virtualenv is already active so all deps will be installed in the
correct site-packages.

Add a `--no-input` flag to prevent being prompted during the sycndb
step. This is mainly for scripted builds.