Lazily set the bundle path, remove `move` method
Using a property handles updating the bundle and config path in one step.
Further, the path does not need to be specified until the setup occurs, so
it `setup` can now take a path. Otherwise a temporary directory is used.

The `move` method has been removed now that setup takes a path argument and
is no longer necessary.