Adds custom-templating functionality
This commit adds the ability to base a Harvest project off of a custom template passed by the -t or --template flag. Additionally, it provides the ability to offload bootstrapping logic to the custom templates own fabric file (as opposed to the commands provided by the Harvest init command). If a passed template contains a Fabric file containing a "harvest_bootstrap" task that task will be run as part of the initilization.