Add support for virtualenvwrapper in Harvest init
Since virtualenvwrapper is in widespread use and fits into the development environment of many python users Harvest should have the ability to create a project and setup the virtualenv according to the conventions of virtualenvwrapper. This allows for consistency and convenience in the workflow of the user.

Instead of relying on virtualenvwrapper's bash script to expose its "mkvirtualenv" utility, we simply reference the WORKON_HOME environment variable to replicate virtualenvwrappers functionality in initially creating the environment. This ensures cross-platform compatibility.