Make fabfile command discovery more lenient.
In order to support custom templating the Harvest init CLI made an attempt to discover available Fabric commands that might be present in the passed template. On some Ubuntu systems this would fail if shell was not specified as `/bin/bash`. Shell is now explicitly stated in this command, furthermore, if the command still fails it will fallback to generic Harvest application initialization. Fixes #31.