Expect full path when template is specified.
Previously the template option expected the URL of a github repository and would build a URL off of that repository assuming that you would like to use the HEAD of that repository as your template. This commit changes the template option so that it expects the full path to a zipball containing a Harvest template. This allows the user to specify a commit, and also allows for the passing of templates not following Github's repository conventions.

Signed-off-by: Tyler Rivera <riverat2@email.chop.edu>